requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autoflake==1.3
Django==2.2.3
django-map-widgets==0.2.2
entrypoints==0.3
epc==0.0.5
flake8==3.7.8
importlib-metadata==0.19
importmagic==0.1.7
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.1.1
pytest==5.0.1
pytest-django==3.5.1
pytz==2019.1
sexpdata==0.0.3
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.2
dj-static
